delete expired session and get pending notification logic

+++ app/src/main/java/estimeet/meetup/model/NotificationEntity.java
@@ -22,6 +22,6 @@ public class NotificationEntity extends BaseModel {
     private int receiverId;
 
     @SerializedName("receiverUId")
-    private int receiverUId;
+    private long receiverUId;
 
 }
